refactor: 테스트 클래스를 테스트 대상 패키지와 동일 패키지로 위치하도록 domain 패키지 추가
