+++ src/test/java/lotto/domain/LotteriesTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.Lotteries;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
