+++ src/test/java/lotto/domain/LottoStatisticsTest.java
@@ -1,8 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.LottoRank;
-import lotto.domain.LottoResult;
-import lotto.domain.LottoStatistics;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
